Добавил где нужно htmlspecialchars, добавил проверку на регитсрацию и обновил логи

+++ app/view/inventory/inventory.php
@@ -137,10 +137,18 @@ if (!isset($points)) {
                                 </td>
                                 <td class="pe-4">
                                     <a href="../defects/defects_view.php?point_id=<?= $point['id'] ?>">Подробнее</a>
-                                    <a href="../../app/view/inventory/update_network_point.php?id=<?php echo htmlspecialchars($point['id']); ?>"
-                                       class="btn btn-outline-secondary btn-sm ms-2">Изменить</a>
-                                    <a href="./delete_network_point.php?id=<?php echo htmlspecialchars($point['id']); ?>"
-                                       class="btn btn-outline-danger btn-sm ms-2" onclick="return confirm('Удалить?')">Удалить</a>
+                                    <?php if (isset($_SESSION['user_info']) && !empty($_SESSION['user_info'])): ?>
+                                        <a href="../../app/view/inventory/update_network_point.php?id=<?php echo htmlspecialchars($point['id']); ?>"
+                                        class="btn btn-outline-secondary btn-sm ms-2">Изменить</a>
+                                        <a href="./delete_network_point.php?id=<?php echo htmlspecialchars($point['id']); ?>"
+                                        class="btn btn-outline-danger btn-sm ms-2" 
+                                        onclick="return confirm('Удалить точку?')">Удалить</a>
+                                    <?php else: ?>
+                                        <button class="btn btn-outline-secondary btn-sm ms-2" disabled 
+                                                title="Доступно только зарегистрированным пользователям">Изменить</button>
+                                        <button class="btn btn-outline-danger btn-sm ms-2" disabled 
+                                                title="Доступно только зарегистрированным пользователям">Удалить</button>
+                                    <?php endif; ?>
                                 </td>
                             </tr>
                         <?php endforeach; ?>
